backend/requirements.txt
# AZURA AQUA Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Config & validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ML & Anomaly detection
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0

# Azure OpenAI
openai>=1.12.0

# OneDrive / Microsoft Graph (optional - for sync from OneDrive)
msal>=1.26.0
httpx>=0.27.0

# Seed script (read REFLEXION.xlsx)
openpyxl>=3.1.0

# Tests
pytest>=7.0.0

# Rate limiting
slowapi>=0.1.9

# PDF generation
reportlab>=4.0.0

# Auth (JWT + password hashing)
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
